feat: cleanup + reorg types
