+++ packages/core/src/index.ts
@@ -1,2 +1,3 @@
 export * from './thread'
+export * from './service'
 export { Block } from './ipld'
